+++ pkg/deployment/reconcile/plan_executor.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2025 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -445,6 +445,8 @@ func (d *Reconciler) executeActionStart(ctx context.Context, action Action) (don
 // createAction create action object based on action type
 func (d *Reconciler) createAction(action api.Action) (Action, ActionContext) {
 	actionCtx := newActionContext(d.log, d.context, &d.metrics)
+
+	actionCtx.SetProgress(action.Progress)
 
 	f, ok := getActionFactory(action.Type)
 	if !ok {