+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - (Feature) UpgradeByReplace Flow
 - (Feature) (Platform) ArangoRoute Timeout option
 - (Feature) Delay Action
+- (Feature) MigrateMember Action
 
 ## [1.2.44](https://github.com/arangodb/kube-arangodb/tree/1.2.44) (2025-02-03)
 - (Maintenance) Kubernetes 1.31.1 libraries